+++ frontend/src/components/mod.rs
@@ -4,4 +4,5 @@ pub mod import_dat_modal;
 pub mod import_rom_modal;
 pub mod navbar;
 pub mod notifications;
+pub mod purge_rom_modal;
 pub mod settings_modal;